requirements.txt
djangorestframework
whitenoise
django-cors-headers